+++ judo/utils/mujoco.py
@@ -2,7 +2,7 @@
 
 import time
 from copy import deepcopy
-from typing import Literal
+from typing import Callable, Literal
 
 import numpy as np
 from mujoco import MjData, MjModel
@@ -20,9 +20,16 @@ def make_model_data_pairs(model: MjModel, num_pairs: int) -> list[tuple[MjModel,
 class RolloutBackend:
     """The backend for conducting multithreaded rollouts."""
 
-    def __init__(self, num_threads: int, backend: Literal["mujoco"]) -> None:
-        """Initialize the backend with a number of threads."""
+    def __init__(self, num_threads: int, backend: Literal["mujoco"], task_to_sim_ctrl: Callable) -> None:
+        """Initialize the backend with a number of threads.
+
+        Args:
+            num_threads: Number of threads for parallel rollout.
+            backend: Backend to use ("mujoco" for Python mujoco rollout).
+            task_to_sim_ctrl: Function to map task controls to simulation controls.
+        """
         self.backend = backend
+        self.task_to_sim_ctrl = task_to_sim_ctrl
         if self.backend == "mujoco":
             self.setup_mujoco_backend(num_threads)
         else:
@@ -57,15 +64,16 @@ class RolloutBackend:
         # shape = (num_rollouts, num_states + 1)
         x0_batched = np.tile(x0, (len(ms), 1))
         full_states = np.concatenate([time.time() * np.ones((len(ms), 1)), x0_batched], axis=-1)
+        processed_controls = self.task_to_sim_ctrl(controls)
         assert full_states.shape[-1] == nq + nv + 1
         assert full_states.ndim == 2
-        assert controls.ndim == 3
-        assert controls.shape[-1] == nu
-        assert controls.shape[0] == full_states.shape[0]
+        assert processed_controls.ndim == 3
+        assert processed_controls.shape[-1] == nu
+        assert processed_controls.shape[0] == full_states.shape[0]
 
         # rollout
         if self.backend == "mujoco":
-            _states, _out_sensors = self.rollout_func(ms, ds, full_states, controls)
+            _states, _out_sensors = self.rollout_func(ms, ds, full_states, processed_controls)
         else:
             raise ValueError(f"Unknown backend: {self.backend}")
         out_states = np.array(_states)[..., 1:]  # remove time from state